Account, AccountRepository, AccountService and the corresponding logic moved out of facilities and into accounts. Cyclic dependency between acounts and facilites resolved.
