+++ accounts/src/main/java/edu/depaul/g6/accounts/repository/AccountRepository.java
@@ -1,6 +1,6 @@
-package edu.depaul.g6.facilities.repository;
+package edu.depaul.g6.accounts.repository;
 
-import edu.depaul.g6.facilities.domain.Account;
+import edu.depaul.g6.accounts.domain.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
