+++ accounts/src/main/java/edu/depaul/g6/accounts/service/Accounts.java
@@ -1,6 +1,8 @@
 package edu.depaul.g6.accounts.service;
 
 //Data
+import edu.depaul.g6.accounts.domain.Account;
+import edu.depaul.g6.accounts.domain.Subscriber;
 import edu.depaul.g6.accounts.repository.UserRepository;
 import edu.depaul.g6.accounts.domain.UserProfile;
 import lombok.Data;
@@ -10,6 +12,7 @@ import org.springframework.stereotype.Service;
 import javax.persistence.*;
 
 import java.io.Serializable;
+import java.security.NoSuchAlgorithmException;
 
 
 @Service
@@ -51,5 +54,26 @@ public class Accounts {
 
     }
      */
+    //TODO This will move out into accounts
+    private SubscriberService subscriberService;
+    private AccountService accountService;
 
+    //TODO This will move out into accounts
+    @Autowired
+    public void setSubscriberService(SubscriberService service) {
+        this.subscriberService = service;
+    }
+    @Autowired void setAccountService(AccountService service) {
+        this.accountService = service;
+    }
+
+    //TODO This will move out into accounts
+    public String saveSubscriber(Subscriber s) throws NoSuchAlgorithmException {
+        return subscriberService.saveSubscriber(s);
+    }
+
+    //TODO This will move out into accounts
+    public Account getAccount(String accountNumber) {
+        return accountService.getAccount(accountNumber);
+    }
 }
